// src/repos/enums/rules.enum.ts
export enum Rules {
    CAN_LIST = 'CAN_LIST',
    CAN_VIEW = 'CAN_VIEW',
    CAN_CREATE = 'CAN_CREATE',
    CAN_EDIT = 'CAN_EDIT',
    CAN_DELETE = 'CAN_DELETE',
    CAN_PURGE = 'CAN_PURGE',
    CAN_ARCHIVE = 'CAN_ARCHIVE',
    CAN_RESTORE = 'CAN_RESTORE',
    CAN_APPROVE = 'CAN_APPROVE',
    CAN_REJECT = 'CAN_REJECT',
    CAN_EXPORT = 'CAN_EXPORT',
    CAN_IMPORT = 'CAN_IMPORT',
    CAN_MANAGE_SETTINGS = 'CAN_MANAGE_SETTINGS',
    CAN_MANAGE_USERS = 'CAN_MANAGE_USERS',
    CAN_ASSIGN_ROLES = 'CAN_ASSIGN_ROLES',
    CAN_ASSIGN_POLICIES = 'CAN_ASSIGN_POLICIES',
    CAN_VIEW_REPORTS = 'CAN_VIEW_REPORTS',
    CAN_GENERATE_REPORTS = 'CAN_GENERATE_REPORTS',
}